+++ src/ts/AdUnits/VastAdUnit.ts
@@ -121,15 +121,6 @@ export class VastAdUnit extends VideoAdUnit<VastCampaign> {
         return this._muted;
     }
 
-    public sendImpressionEvent(sessionId: string, sdkVersion: number): void {
-        const impressionUrls = this._vastCampaign.getVast().getImpressionUrls();
-        if (impressionUrls) {
-            for (const impressionUrl of impressionUrls) {
-                this.sendThirdPartyEvent('vast impression', sessionId, sdkVersion, impressionUrl);
-            }
-        }
-    }
-
     public sendTrackingEvent(eventName: string, sessionId: string, sdkVersion: number): void {
         const trackingEventUrls = this._vastCampaign.getVast().getTrackingEventUrls(eventName);
         if (trackingEventUrls) {
@@ -137,12 +128,6 @@ export class VastAdUnit extends VideoAdUnit<VastCampaign> {
                 this.sendThirdPartyEvent(`vast ${eventName}`, sessionId, sdkVersion, url);
             }
         }
-    }
-
-    public sendProgressEvents(sessionId: string, sdkVersion: number, position: number, oldPosition: number) {
-        this.sendQuartileEvent(sessionId, sdkVersion, position, oldPosition, 1, 'firstQuartile');
-        this.sendQuartileEvent(sessionId, sdkVersion, position, oldPosition, 2, 'midpoint');
-        this.sendQuartileEvent(sessionId, sdkVersion, position, oldPosition, 3, 'thirdQuartile');
     }
 
     public getVideoClickThroughURL(): string | null {
@@ -219,23 +204,10 @@ export class VastAdUnit extends VideoAdUnit<VastCampaign> {
         }
     }
 
-    private sendQuartileEvent(sessionId: string, sdkVersion: number, position: number, oldPosition: number, quartile: number, quartileEventName: string) {
-        if (this.getTrackingEventUrls(quartileEventName)) {
-            const duration = this._vastCampaign.getVideo().getDuration();
-            if (duration && duration > 0 && position > duration * 0.25 * quartile && oldPosition < duration * 0.25 * quartile) {
-                this.sendTrackingEvent(quartileEventName, sessionId, sdkVersion);
-            }
-        }
-    }
-
     private sendThirdPartyEvent(event: string, sessionId: string, sdkVersion: number, url: string): void {
         url = url.replace(/%ZONE%/, this._vastPlacement.getId());
         url = url.replace(/%SDK_VERSION%/, sdkVersion.toString());
         this._thirdPartyEventManager.sendEvent(event, sessionId, url, this._vastCampaign.getUseWebViewUserAgentForTracking());
-    }
-
-    private getTrackingEventUrls(eventName: string): string[] | null {
-        return this._vastCampaign.getVast().getTrackingEventUrls(eventName);
     }
 
     private isValidURL(url: string | null): boolean {