+++ src/ts/Managers/XPromoOperativeEventManager.ts
@@ -1,9 +1,10 @@
 import { IOperativeEventManagerParams, OperativeEventManager } from 'Managers/OperativeEventManager';
-import { Session } from 'Models/Session';
-import { AdUnitStyle } from 'Models/AdUnitStyle';
 import { Placement } from 'Models/Placement';
-import { IXPromoCampaign, XPromoCampaign } from 'Models/Campaigns/XPromoCampaign';
+import { XPromoCampaign } from 'Models/Campaigns/XPromoCampaign';
 import { HttpKafka } from 'Utilities/HttpKafka';
+import { PlayerMetaData } from 'Models/MetaData/PlayerMetaData';
+import { EventType } from 'Models/Session';
+import { INativeResponse } from 'Utilities/Request';
 
 export class XPromoOperativeEventManager extends OperativeEventManager {
     private _xPromoCampaign: XPromoCampaign;
@@ -14,7 +15,54 @@ export class XPromoOperativeEventManager extends OperativeEventManager {
         this._xPromoCampaign = params.campaign;
     }
 
-    public sendHttpKafkaEvent(kafkaType: string, eventType: string, placement: Placement, videoOrientation?: string) {
+    public sendStartEvent(placement: Placement, videoOrientation?: string): Promise<void> {
+        const session = this._campaign.getSession();
+
+        if(session.getEventSent(EventType.START)) {
+            return Promise.resolve();
+        }
+
+        session.setEventSent(EventType.START);
+
+        return this._metaDataManager.fetch(PlayerMetaData, false).then(player => {
+            if(player) {
+                this.setGamerServerId(player.getServerId());
+            } else {
+                this.setGamerServerId(undefined);
+            }
+            return this.sendHttpKafkaEvent('ads.xpromo.operative.videostart.v1.json', 'start', placement, videoOrientation);
+        }).then(() => {
+            return;
+        });
+    }
+
+    public sendViewEvent(placement: Placement, videoOrientation?: string): Promise<void> {
+        const session = this._campaign.getSession();
+
+        if(session.getEventSent(EventType.VIEW)) {
+            return Promise.resolve(void(0));
+        }
+        session.setEventSent(EventType.VIEW);
+
+        return this.sendHttpKafkaEvent('ads.xpromo.operative.videoview.v1.json', 'view', placement, videoOrientation).then(() => {
+            return;
+        });
+    }
+
+    public sendClickEvent(placement: Placement, videoOrientation?: string): Promise<void> {
+        const session = this._campaign.getSession();
+
+        if(session.getEventSent(EventType.CLICK)) {
+            return Promise.resolve(void(0));
+        }
+        session.setEventSent(EventType.CLICK);
+
+        return this.sendHttpKafkaEvent('ads.xpromo.operative.videoclick.v1.json', 'click', placement, videoOrientation).then(() => {
+            return;
+        });
+    }
+
+    public sendHttpKafkaEvent(kafkaType: string, eventType: string, placement: Placement, videoOrientation?: string): Promise<INativeResponse> {
         const fulfilled = ([id, infoJson]: [string, any]) => {
 
             // todo: clears duplicate data for httpkafka, should be cleaned up
@@ -37,7 +85,7 @@ export class XPromoOperativeEventManager extends OperativeEventManager {
             infoJson.sourceGameId = this._clientInfo.getGameId();
             infoJson.targetGameId = this._xPromoCampaign.getGameId().toString();
 
-            HttpKafka.sendEvent(kafkaType, infoJson);
+            return HttpKafka.sendEvent(kafkaType, infoJson);
         };
 
         return this.createUniqueEventMetadata(placement, this._sessionManager.getGameSessionId(), this._gamerServerId, OperativeEventManager.getPreviousPlacementId(), videoOrientation).then(fulfilled);